Finish daily repeat task

+++ app/src/main/java/us/bojie/a_great_day/data/firebase/FirebaseManager.kt
@@ -43,8 +43,10 @@ class FirebaseManager @Inject constructor(
         }
     }
 
-    suspend fun updateTaskToNextDay(task: Task): Boolean {
-        deleteTask(task)
+    suspend fun updateTaskToNextDay(task: Task, needDelete: Boolean = true): Boolean {
+        if (needDelete) {
+            deleteTask(task)
+        }
         return updateTask(task, formattedNextDay)
     }
